+++ deno/payloads/v10/_interactions/_applicationCommands/_chatInput/attachment.ts
@@ -2,8 +2,11 @@ import type { Snowflake } from '../../../../../globals.ts';
 import type { APIApplicationCommandOptionBase, APIInteractionDataOptionBase } from './base.ts';
 import type { ApplicationCommandOptionType } from './shared.ts';
 
-export type APIApplicationCommandAttachmentOption =
-	APIApplicationCommandOptionBase<ApplicationCommandOptionType.Attachment>;
+export type FileUploadType = 'audio' | 'image' | 'video' | `.${string}`;
+
+export interface APIApplicationCommandAttachmentOption extends APIApplicationCommandOptionBase<ApplicationCommandOptionType.Attachment> {
+	file_types?: FileUploadType[];
+}
 
 export type APIApplicationCommandInteractionDataAttachmentOption = APIInteractionDataOptionBase<
 	ApplicationCommandOptionType.Attachment,