+++ payloads/v9/message.ts
@@ -5,7 +5,12 @@ import type { _NonNullableFields } from '../../utils/internals';
 import type { APIApplication } from './application';
 import type { APIChannel, APIThreadChannel, APIThreadMember, ChannelType } from './channel';
 import type { APIPartialEmoji } from './emoji';
-import type { APIInteractionDataResolved, APIMessageInteraction, APIMessageInteractionMetadata } from './interactions';
+import type {
+	APIInteractionDataResolved,
+	APIMessageInteraction,
+	APIMessageInteractionMetadata,
+	FileUploadType,
+} from './interactions';
 import type { APIRole } from './permissions';
 import type { APIPoll } from './poll';
 import type { APISticker, APIStickerItem } from './sticker';
@@ -143,7 +148,7 @@ export interface APIBaseMessageNoChannel {
 	/**
 	 * Sent with Rich Presence-related chat embeds
 	 *
-	 * @see {@link https://discord.com/developers/docs/resources/application#application-object}
+	 * @see {@link https://docs.discord.com/developers/resources/application#application-object}
 	 */
 	application?: Partial<APIApplication>;
 	/**
@@ -368,6 +373,7 @@ export enum MessageActivityType {
 	Spectate,
 	Listen,
 	JoinRequest = 5,
+	StreamRequest,
 }
 
 /**
@@ -959,7 +965,7 @@ export interface APIAttachment {
 	 */
 	ephemeral?: boolean;
 	/**
-	 * The duration of the audio file (currently for voice messages)
+	 * The duration of the audio or video file
 	 */
 	duration_secs?: number;
 	/**
@@ -1290,9 +1296,7 @@ export interface APIButtonComponentWithSKUId extends APIButtonBase<ButtonStyle.P
  * @see {@link https://discord.com/developers/docs/components/reference#button}
  */
 export type APIButtonComponent =
-	| APIButtonComponentWithCustomId
-	| APIButtonComponentWithSKUId
-	| APIButtonComponentWithURL;
+	APIButtonComponentWithCustomId | APIButtonComponentWithSKUId | APIButtonComponentWithURL;
 
 /**
  * @see {@link https://discord.com/developers/docs/components/reference#button-button-styles}
@@ -1498,10 +1502,7 @@ export interface APISelectMenuDefaultValue<T extends SelectMenuDefaultValueType>
  * @see {@link https://discord.com/developers/docs/components/reference}
  */
 export type APIAutoPopulatedSelectMenuComponent =
-	| APIChannelSelectComponent
-	| APIMentionableSelectComponent
-	| APIRoleSelectComponent
-	| APIUserSelectComponent;
+	APIChannelSelectComponent | APIMentionableSelectComponent | APIRoleSelectComponent | APIUserSelectComponent;
 
 /**
  * @see {@link https://discord.com/developers/docs/components/reference}
@@ -1885,6 +1886,15 @@ export interface APIFileUploadComponent extends APIBaseComponent<ComponentType.F
 	 * Maximum number of items that can be uploaded (defaults to 1); max 10
 	 */
 	max_values?: number;
+	/**
+	 * Allowed file types that can be uploaded; max 10
+	 *
+	 * If only dot-prefixed extensions are specified, you must also include `.jpg` for image uploads
+	 * and both `.mp4` and `.mov` for video uploads because of mobile platform limitations.
+	 *
+	 * @see {@link https://docs.discord.com/developers/reference#file-type-filtering}
+	 */
+	file_types?: FileUploadType[];
 	/**
 	 * Whether the file upload requires files to be uploaded before submitting the modal
 	 *